From http://plugins.jquery.com/node/3955
September 11, 2008 - 12:16pm - sc0ttkclark
Functionality on textareas with the width:100% style property is
erratic. Happens no matter what I try unless I change it to width:XXpx

January 5, 2009 - 2:18am - karex
To be able to autogrow relatively sized textareas, replace the
following line (it's the 99th):
'width'      : this.textarea.css('width'),
to this:
'width'      : this.textarea.width() + 'px',

Note: the textarea in question must already be part of the DOM to
measure its width.